IntValue StringValue BoolValue, ValueEvent not showing up in unity editor

+++ src/Id/Components/IntValue.cs
@@ -4,7 +4,7 @@ using System.Collections.Generic;
 
 namespace MicroState.Id.Components
 {
-	public class FloatValue : SingleValueBase<float>
+	public class IntValue : SingleValueBase<int>
 	{
 	}
 }
